[WNMGDS-716] Convert TextField and Select to typescript (#1050)
* Convert TextField to tsx

* Convert Mask to tsx

* First pass of changing Select and TextField to typescript

* Fix bugs

* lint

* Fix errors and missing props

* Fix ref

* Update snapshots with default type

* Update index.d.ts

* Fix mgov errors
