+++ packages/design-system-docs/src/pages/components/Dropdown/_Dropdown.docs.scss
@@ -19,7 +19,7 @@ Class-based component gives flexibility for active focus management by allowing 
 
 @react-example Dropdown.example.jsx
 
-@react-props Dropdown.jsx
+@react-props Dropdown.tsx
 
 Style guide: components.dropdown.react
 */
