+++ packages/design-system-docs/src/pages/components/TextField/_TextField.docs.scss
@@ -17,7 +17,7 @@ A `TextField` component renders an input field as well as supporting UI elements
 
 @react-example TextField.example.jsx
 
-@react-props TextField.jsx
+@react-props TextField.tsx
 
 Style guide: components.text-field.react
 */
