+++ packages/design-system/src/components/Dropdown/Select.tsx
@@ -1,14 +1,92 @@
-import PropTypes from 'prop-types';
 import React from 'react';
 import classNames from 'classnames';
+
+export type SelectDefaultValue = number | string;
+
+export interface SelectOptions {
+  label: React.ReactNode;
+  value: number | string;
+}
+
+export type SelectSize = 'small' | 'medium';
+
+export type SelectValue = number | string;
+
+export type SelectErrorPlacement = 'top' | 'bottom';
+
+export interface SelectProps {
+  /**
+   * Adds `aria-label` attribute. When using `aria-label`, `label` should be empty string.
+   */
+  ariaLabel?: string;
+  /**
+   * Used to define custom Select options (i.e. option groups). When using the `children` prop, `options` should be an empty list.
+   */
+  children?: React.ReactNode;
+  /**
+   * Sets the initial selected state. Use this for an uncontrolled component;
+   * otherwise, use the `value` property.
+   */
+  defaultValue?: SelectDefaultValue;
+  /**
+   * Disables the entire field.
+   */
+  disabled?: boolean;
+  /**
+   * The ID of the error message applied to the Select field.
+   */
+  errorId?: string;
+  errorMessage?: React.ReactNode;
+  /**
+   * Location of the error message relative to the field input
+   */
+  errorPlacement: SelectErrorPlacement;
+  /**
+   * Additional classes to be added to the select element
+   */
+  fieldClassName?: string;
+  /**
+   * A unique ID to be used for the Select field.
+   */
+  id: string;
+  /**
+   * Applies the "inverse" UI theme
+   */
+  inversed?: boolean;
+  /**
+   * The field's `name` attribute
+   */
+  name: string;
+  /**
+   * The list of options to be rendered. Provide an empty list if using custom options via the `children` prop.
+   */
+  options: SelectOptions[];
+  onBlur?: (...args: any[]) => any;
+  onChange?: (...args: any[]) => any;
+  setRef?: (...args: any[]) => any;
+  /**
+   * If the component renders a select, set the max-width of the input either to `'small'` or `'medium'`.
+   */
+  size?: SelectSize;
+  /**
+   * Sets the field's `value`. Use this in combination with `onChange`
+   * for a controlled component; otherwise, set `defaultValue`.
+   */
+  value?: SelectValue;
+}
+
+type OmitProps = 'size' | 'value';
 
 /**
  * <Select> is an internal component used by <Dropdown>, which wraps it and handles common form controls like labels, error messages, etc
  * <Select> is also exported for advanced design system use cases, where the internal component can be leveraged to build custom form components
  * As an internal component, it's subject to more breaking changes. Exercise caution using <Select> outside of those special cases
  */
-export class Select extends React.PureComponent {
-  constructor(props) {
+export class Select extends React.PureComponent<
+  Omit<React.ComponentPropsWithoutRef<'select'>, OmitProps> & SelectProps,
+  any
+> {
+  constructor(props: SelectProps) {
     super(props);
 
     if (process.env.NODE_ENV !== 'production') {
@@ -20,7 +98,7 @@ export class Select extends React.PureComponent {
     }
   }
 
-  render() {
+  render(): React.ReactNode {
     // Select specific props
     const {
       ariaLabel,
@@ -72,71 +150,5 @@ export class Select extends React.PureComponent {
     );
   }
 }
-
-Select.propTypes = {
-  /**
-   * Adds `aria-label` attribute. When using `aria-label`, `label` should be empty string.
-   */
-  ariaLabel: PropTypes.string,
-  /**
-   * Used to define custom Select options (i.e. option groups). When using the `children` prop, `options` should be an empty list.
-   */
-  children: PropTypes.node,
-  /**
-   * Sets the initial selected state. Use this for an uncontrolled component;
-   * otherwise, use the `value` property.
-   */
-  defaultValue: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
-  /**
-   * Disables the entire field.
-   */
-  disabled: PropTypes.bool,
-  /**
-   * The ID of the error message applied to the Select field.
-   */
-  errorId: PropTypes.string,
-  errorMessage: PropTypes.node,
-  /**
-   * Location of the error message relative to the field input
-   */
-  errorPlacement: PropTypes.oneOf(['top', 'bottom']).isRequired,
-  /**
-   * Additional classes to be added to the select element
-   */
-  fieldClassName: PropTypes.string,
-  /**
-   * A unique ID to be used for the Select field.
-   */
-  id: PropTypes.string.isRequired,
-  /**
-   * Applies the "inverse" UI theme
-   */
-  inversed: PropTypes.bool,
-  /**
-   * The field's `name` attribute
-   */
-  name: PropTypes.string.isRequired,
-  /**
-   * The list of options to be rendered. Provide an empty list if using custom options via the `children` prop.
-   */
-  options: PropTypes.arrayOf(
-    PropTypes.shape({
-      label: PropTypes.node.isRequired,
-      value: PropTypes.oneOfType([PropTypes.number, PropTypes.string]).isRequired,
-    })
-  ).isRequired,
-  onBlur: PropTypes.func,
-  onChange: PropTypes.func,
-  setRef: PropTypes.func,
-  /**
-   * If the component renders a select, set the max-width of the input either to `'small'` or `'medium'`.
-   */
-  size: PropTypes.oneOf(['small', 'medium']),
-  /**
-   * Sets the field's `value`. Use this in combination with `onChange`
-   * for a controlled component; otherwise, set `defaultValue`.
-   */
-  value: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
-};
 
 export default Select;
