+++ packages/design-system/src/types/index.d.ts
@@ -6,8 +6,6 @@ export { default as Autocomplete } from './Autocomplete/Autocomplete';
 export { default as Badge } from './Badge/Badge';
 export { default as Button } from './Button/Button';
 export { default as Dialog } from './Dialog/Dialog';
-export { default as Dropdown } from './Dropdown/Dropdown';
-export { default as Select } from './Dropdown/Select';
 export { default as FormLabel } from './FormLabel/FormLabel';
 export { default as HelpDrawer } from './HelpDrawer/HelpDrawer';
 export { default as HelpDrawerToggle } from './HelpDrawer/HelpDrawerToggle';
@@ -24,9 +22,6 @@ export { default as TableHead } from './Table/TableHead';
 export { default as TableRow } from './Table/TableRow';
 export { default as Tabs } from './Tabs/Tabs';
 export { default as TabPanel } from './Tabs/TabPanel';
-export { default as TextField } from './TextField/TextField';
-export { default as TextInput } from './TextField/TextInput';
-export { maskValue, unmaskValue } from './TextField/Mask';
 export { default as Tooltip } from './Tooltip/Tooltip';
 export { default as TooltipIcon } from './Tooltip/TooltipIcon';
 export { default as UsaBanner } from './UsaBanner/UsaBanner';
@@ -40,7 +35,12 @@ export * from './ChoiceList/Choice';
 export * from './ChoiceList/ChoiceList';
 export * from './DateField/DateField';
 export * from './DateField/DateInput';
+export * from './Dropdown/Dropdown';
+export * from './Dropdown/Select';
 export * from './InlineError/InlineError';
 export * from './FilterChip/FilterChip';
 export * from './FormControl/FormControl';
+export * from './TextField/Textfield';
+export * from './TextField/TextInput';
+export * from './Textfield/Mask';
 export * from './flags';
